# Compiler: don't always use Array for node dependencies and observers