--- a/src/compiler/crystal/interpreter/multidispatch.cr
+++ b/src/compiler/crystal/interpreter/multidispatch.cr
@@ -39,7 +39,7 @@ require "../semantic/main_visitor"
 # end
 # ```
 module Crystal::Repl::Multidispatch
-  def self.create_def(context : Context, node : Call, target_defs : Array(Def))
+  def self.create_def(context : Context, node : Call, target_defs : ZeroOneOrMany(Def))
     if node.block
       a_def = create_def_uncached(context, node, target_defs)
 
@@ -70,7 +70,7 @@ module Crystal::Repl::Multidispatch
     a_def
   end
 
-  private def self.create_def_uncached(context : Context, node : Call, target_defs : Array(Def))
+  private def self.create_def_uncached(context : Context, node : Call, target_defs : ZeroOneOrMany(Def))
     autocast_types = nil
 
     # The generated multidispatch method should handle autocasted
@@ -200,7 +200,7 @@ module Crystal::Repl::Multidispatch
         end
 
       call = Call.new(call_obj, node.name, call_args)
-      call.target_defs = [target_def]
+      call.target_defs = ZeroOneOrMany(Def).new(target_def)
       call.type = target_def.type
 
       if block