--- a/src/compiler/crystal/semantic/bindings.cr
+++ b/src/compiler/crystal/semantic/bindings.cr
@@ -111,7 +111,7 @@ module Crystal
       end
     end
 
-    def bind_to(nodes : Array)
+    def bind_to(nodes : Enumerable(ASTNode))
       return if nodes.empty?
 
       bind do