--- a/src/compiler/crystal/semantic/bindings.cr
+++ b/src/compiler/crystal/semantic/bindings.cr
@@ -108,19 +108,9 @@ module Crystal
 
     def bind_to(node : ASTNode)
       bind(node) do
-        @dependencies += node
+        @dependencies.push node
         node.add_observer self
         node
-      end
-    end
-
-    def bind_to(node1 : ASTNode, node2 : ASTNode)
-      bind do
-        @dependencies += node1
-        @dependencies += node2
-        node1.add_observer self
-        node2.add_observer self
-        node1
       end
     end
 
@@ -128,7 +118,7 @@ module Crystal
       return if nodes.empty?
 
       bind do
-        @dependencies += nodes
+        @dependencies.concat nodes
         nodes.each &.add_observer self
         nodes.first
       end
@@ -160,12 +150,12 @@ module Crystal
     end
 
     def unbind_from(node : ASTNode) : Nil
-      @dependencies = @dependencies.reject &.same?(node)
+      @dependencies.reject! &.same?(node)
       node.remove_observer self
     end
 
     def unbind_from(nodes : Enumerable(ASTNode)) : Nil
-      @dependencies = @dependencies.reject { |dependency|
+      @dependencies.reject! { |dependency|
         nodes.any? &.same?(dependency)
       }
       nodes.each &.remove_observer self
@@ -184,11 +174,11 @@ module Crystal
     end
 
     def add_observer(observer : ASTNode) : Nil
-      @observers += observer
+      @observers.push observer
     end
 
     def remove_observer(observer : ASTNode) : Nil
-      @observers = @observers.reject &.same?(observer)
+      @observers.reject! &.same?(observer)
     end
 
     def set_enclosing_call(enclosing_call)