… spec/compiler/crystal/zero_one_or_many_spec.cr
… src/compiler/crystal/semantic/cleanup_transformer.cr
… src/compiler/crystal/semantic/cover.cr
… src/compiler/crystal/semantic/filters.cr
… src/compiler/crystal/semantic/fix_missing_types.cr
… src/compiler/crystal/semantic/lib.cr
… src/compiler/crystal/semantic/literal_expander.cr
… src/compiler/crystal/semantic/main_visitor.cr
… src/compiler/crystal/semantic/match.cr
… src/compiler/crystal/semantic/method_lookup.cr
… src/compiler/crystal/semantic/warnings.cr
… src/compiler/crystal/tools/context.cr
… src/compiler/crystal/tools/implementations.cr
… src/compiler/crystal/zero_one_or_many.cr