This PR changes how bindings and observers are stored in code.

## What are bindings and observers?

When you have code like this:

```crystal
a = 1
if rand < 0.5
  b = a
else
  b = 0
end
```

What Crystal does is to bind nodes between each other to propagate types. In the example above:
- `a` is bound to `1`
- `b` is bound to `a` in one branch, and to `0` in another

The example above is very simple, but the idea of binding is that if a node's type changes (and that can happen in Crystal if we are inside a loop) then the types are recomputed across nodes.

These are stored bi-directionally:
- dependencies is used to compute the type of a node. In this case the dependencies of `b` are `a` and `0`
- observers are used to know which nodes to propagate types in case a node's type changes. For example the observer of `a` is `b`

We store dependencies and observers as arrays.

## What did I find

It turns out, most of the time the dependencies and observers arrays only have one element. For example, when compiling the compiler, we get this:

```
dependencies 
------------
array size: occurrences
1: 1278339
2: 137905
3: 6840
4: 1272
5: 644
6: 785
7: 1574
8: 263
...

observers
---------
array size: occurrences
1: 589519
2: 54349
3: 123616
4: 6272
5: 1004
6: 331
7: 162
8: 62
```

That means we are allocating a lot of arrays of size 1 when we could avoid allocating all that memory. Allocating memory well, requires memory, and also puts more pressure into the GC.

## What's the change

We change `dependencies` and `observers` from `Array(ASTNode)?` to `Nil | ASTNode | Array(ASTNode)`. The middle type is the case of a single bound node. It slightly complicates the compiler's logic, but not that much.

## Benchmark

This is compiling the compiler with the old compiler, no warm cache:

```
✗ time bin/crystal build -s src/compiler/crystal.cr
Parse:                             00:00:00.000038250 (   0.83MB)
Semantic (top level):              00:00:00.317475667 ( 143.70MB)
Semantic (new):                    00:00:00.002673166 ( 143.70MB)
Semantic (type declarations):      00:00:00.040537083 ( 143.70MB)
Semantic (abstract def check):     00:00:00.015006458 ( 159.70MB)
Semantic (restrictions augmenter): 00:00:00.009682959 ( 159.70MB)
Semantic (ivars initializers):     00:00:04.800375875 (1815.61MB)
Semantic (cvars initializers):     00:00:00.005825667 (1815.61MB)
Semantic (main):                   00:00:09.014618542 (2111.61MB)
Semantic (cleanup):                00:00:00.000723083 (2111.61MB)
Semantic (recursive struct check): 00:00:00.001276250 (2111.61MB)
Codegen (crystal):                 00:00:03.229728375 (2319.61MB)
Codegen (bc+obj):                  00:00:04.390755792 (2319.61MB)
Codegen (linking):                 00:00:00.644151084 (2319.61MB)
dsymutil:                          00:00:00.382296125 (2319.61MB)

bin/crystal build -s src/compiler/crystal.cr  38.90s user 2.33s system 179% cpu 22.973 total
```

With a warm cache:

```
$ time bin/crystal build -s src/compiler/crystal.cr
Parse:                             00:00:00.000038417 (   0.83MB)
Semantic (top level):              00:00:00.315096292 ( 143.70MB)
Semantic (new):                    00:00:00.002657209 ( 143.70MB)
Semantic (type declarations):      00:00:00.041264208 ( 143.70MB)
Semantic (abstract def check):     00:00:00.015403083 ( 159.70MB)
Semantic (restrictions augmenter): 00:00:00.009669917 ( 159.70MB)
Semantic (ivars initializers):     00:00:04.605003917 (1815.61MB)
Semantic (cvars initializers):     00:00:00.005939417 (1815.61MB)
Semantic (main):                   00:00:00.824864833 (2023.61MB)
Semantic (cleanup):                00:00:00.000550917 (2023.61MB)
Semantic (recursive struct check): 00:00:00.001286000 (2023.61MB)
Codegen (crystal):                 00:00:03.887214208 (2303.61MB)
Codegen (bc+obj):                  00:00:00.500315541 (2303.61MB)
Codegen (linking):                 00:00:00.636580958 (2303.61MB)
dsymutil:                          00:00:00.378992750 (2303.61MB)
bin/crystal build -s src/compiler/crystal.cr  17.08s user 1.44s system 163% cpu 11.342 total
```

So around 2.3GB of memory, 22.9s without cache and 11.34s with cache.

With this PR, compiling the compiler, no warm cache:

```
$ time bin/crystal build -s src/compiler/crystal.cr
Using compiled compiler at .build/crystal
Parse:                             00:00:00.000299458 (   0.83MB)
Semantic (top level):              00:00:00.321059583 ( 143.78MB)
Semantic (new):                    00:00:00.002321625 ( 143.78MB)
Semantic (type declarations):      00:00:00.039610125 ( 143.78MB)
Semantic (abstract def check):     00:00:00.014762666 ( 159.78MB)
Semantic (restrictions augmenter): 00:00:00.009603708 ( 159.78MB)
Semantic (ivars initializers):     00:00:04.492760667 (1559.70MB)
Semantic (cvars initializers):     00:00:00.005612375 (1559.70MB)
Semantic (main):                   00:00:07.867504833 (1807.70MB)
Semantic (cleanup):                00:00:00.000605042 (1807.70MB)
Semantic (recursive struct check): 00:00:00.001187625 (1807.70MB)
Codegen (crystal):                 00:00:03.365013917 (2111.70MB)
Codegen (bc+obj):                  00:00:04.369402625 (2111.70MB)
Codegen (linking):                 00:00:00.651541375 (2111.70MB)
dsymutil:                          00:00:00.374684958 (2111.70MB)
bin/crystal build -s src/compiler/crystal.cr  37.83s user 2.28s system 174% cpu 22.933 total
```

With a warm cache:

```
$ time bin/crystal build -s src/compiler/crystal.cr
Using compiled compiler at .build/crystal
Parse:                             00:00:00.000036541 (   0.83MB)
Semantic (top level):              00:00:00.312257667 ( 143.78MB)
Semantic (new):                    00:00:00.002320750 ( 143.78MB)
Semantic (type declarations):      00:00:00.039073250 ( 143.78MB)
Semantic (abstract def check):     00:00:00.014990125 ( 159.78MB)
Semantic (restrictions augmenter): 00:00:00.009610959 ( 159.78MB)
Semantic (ivars initializers):     00:00:04.531297042 (1559.70MB)
Semantic (cvars initializers):     00:00:00.005627000 (1559.70MB)
Semantic (main):                   00:00:00.801419375 (1735.70MB)
Semantic (cleanup):                00:00:00.000553834 (1735.70MB)
Semantic (recursive struct check): 00:00:00.001289375 (1735.70MB)
Codegen (crystal):                 00:00:03.953488458 (2111.70MB)
Codegen (bc+obj):                  00:00:00.495159166 (2111.70MB)
Codegen (linking):                 00:00:00.632672083 (2111.70MB)
dsymutil:                          00:00:00.376506500 (2111.70MB)
bin/crystal build -s src/compiler/crystal.cr  16.43s user 1.42s system 158% cpu 11.293 total
```

So 2.111GB of memory, and about the same time.

I hoped the times would go down a bit more, but at least 200MB less for compiling the compiler seems good.