@asterite 2022-08-19
Note: I'm still not convinced we should merge this.

Maybe others can run benchmarks on their machines, with their projects, to see if there's any noticeable difference.