@jwoertink 2022-08-19
For fun I decided to give this a try:

This is compiling my app with the compiler on master (1.6.0.dev), no warm cache:

```
❯ time /home/jeremy/Development/crystal/lang/bin/crystal build -s src/start_server.cr -o bin/start_server

Parse:                             00:00:00.000037160 (   0.76MB)
Semantic (top level):              00:00:21.296049102 ( 652.70MB)
Semantic (new):                    00:00:00.006644233 ( 652.70MB)
Semantic (type declarations):      00:00:00.115471755 ( 652.70MB)
Semantic (abstract def check):     00:00:00.036184495 ( 684.70MB)
Semantic (restrictions augmenter): 00:00:00.030057488 ( 700.70MB)
Semantic (ivars initializers):     00:00:00.132122853 ( 780.70MB)
Semantic (cvars initializers):     00:00:00.018864900 ( 796.70MB)
Semantic (main):                   00:00:07.086655243 (2283.70MB)
Semantic (cleanup):                00:00:00.001547965 (2283.70MB)
Semantic (recursive struct check): 00:00:00.003476881 (2283.70MB)
Codegen (crystal):                 00:00:07.836129589 (3019.70MB)
Codegen (bc+obj):                  00:00:08.676941735 (3019.70MB)
Codegen (linking):                 00:00:04.986041171 (3019.70MB)

Macro runs:
 - /home/jeremy/Sites/joysticktv/lib/teeplate/src/lib/file_tree/macros/directory.cr: 00:00:09.341495853
 - /home/jeremy/Sites/joysticktv/lib/lucky/src/run_macros/generate_asset_helpers.cr: 00:00:10.235055927
 - /home/jeremy/Development/crystal/lang/src/ecr/process.cr: reused previous compilation (00:00:00.002888857)
real	0m50.348s
user	1m21.199s
sys	0m5.408s
```

With a warm cache:

```
❯ time /home/jeremy/Development/crystal/lang/bin/crystal build -s src/start_server.cr -o bin/start_server

Parse:                             00:00:00.000027830 (   0.76MB)
Semantic (top level):              00:00:01.079584170 ( 555.76MB)
Semantic (new):                    00:00:00.007232593 ( 555.76MB)
Semantic (type declarations):      00:00:00.112843284 ( 555.76MB)
Semantic (abstract def check):     00:00:00.034558802 ( 587.76MB)
Semantic (restrictions augmenter): 00:00:00.029456914 ( 587.76MB)
Semantic (ivars initializers):     00:00:00.133077860 ( 683.76MB)
Semantic (cvars initializers):     00:00:00.018820298 ( 683.76MB)
Semantic (main):                   00:00:06.719366274 (2124.69MB)
Semantic (cleanup):                00:00:00.001525072 (2124.69MB)
Semantic (recursive struct check): 00:00:00.003323244 (2124.69MB)
Codegen (crystal):                 00:00:10.018644245 (2815.69MB)
Codegen (bc+obj):                  00:00:04.566887130 (2815.69MB)
Codegen (linking):                 00:00:04.789592386 (2815.69MB)

Macro runs:
 - /home/jeremy/Sites/joysticktv/lib/teeplate/src/lib/file_tree/macros/directory.cr: reused previous compilation (00:00:00.002541069)
 - /home/jeremy/Sites/joysticktv/lib/lucky/src/run_macros/generate_asset_helpers.cr: reused previous compilation (00:00:00.002758590)
 - /home/jeremy/Development/crystal/lang/src/ecr/process.cr: reused previous compilation (00:00:00.002699890)

Codegen (bc+obj):
 - 4269/4412 .o files were reused

These modules were not reused:
 - a very long list of my code lol

real	0m27.632s
user	0m35.284s
sys	0m4.272s
```

This branch with no warm cache:

```
❯ time /home/jeremy/Development/crystal/lang/bin/crystal build -s src/start_server.cr -o bin/start_server

Parse:                             00:00:00.000033191 (   0.76MB)
Semantic (top level):              00:00:01.111511339 ( 554.45MB)
Semantic (new):                    00:00:00.007689575 ( 554.45MB)
Semantic (type declarations):      00:00:00.114648171 ( 554.45MB)
Semantic (abstract def check):     00:00:00.033110647 ( 570.45MB)
Semantic (restrictions augmenter): 00:00:00.029217100 ( 586.45MB)
Semantic (ivars initializers):     00:00:00.131192539 ( 666.45MB)
Semantic (cvars initializers):     00:00:00.018006393 ( 666.45MB)
Semantic (main):                   00:00:07.057628712 (1851.38MB)
Semantic (cleanup):                00:00:00.001507374 (1851.38MB)
Semantic (recursive struct check): 00:00:00.003049429 (1851.38MB)
Codegen (crystal):                 00:00:09.843128251 (2574.38MB)
Codegen (bc+obj):                  00:00:04.726382542 (2574.38MB)
Codegen (linking):                 00:00:04.785858078 (2574.38MB)

Macro runs:
 - /home/jeremy/Sites/joysticktv/lib/teeplate/src/lib/file_tree/macros/directory.cr: reused previous compilation (00:00:00.002809027)
 - /home/jeremy/Sites/joysticktv/lib/lucky/src/run_macros/generate_asset_helpers.cr: reused previous compilation (00:00:00.002745616)
 - /home/jeremy/Development/crystal/lang/src/ecr/process.cr: reused previous compilation (00:00:00.002840708)
real	0m27.988s
user	0m32.999s
sys	0m4.424s
```

This branch with warm cache:

```
❯ time /home/jeremy/Development/crystal/lang/bin/crystal build -s src/start_server.cr -o bin/start_server

Parse:                             00:00:00.000030140 (   0.76MB)
Semantic (top level):              00:00:01.083887540 ( 555.63MB)
Semantic (new):                    00:00:00.007279427 ( 555.63MB)
Semantic (type declarations):      00:00:00.114083008 ( 555.63MB)
Semantic (abstract def check):     00:00:00.033601663 ( 571.63MB)
Semantic (restrictions augmenter): 00:00:00.028796460 ( 587.63MB)
Semantic (ivars initializers):     00:00:00.128038875 ( 667.63MB)
Semantic (cvars initializers):     00:00:00.017660649 ( 667.63MB)
Semantic (main):                   00:00:06.503724956 (1852.56MB)
Semantic (cleanup):                00:00:00.001476434 (1852.56MB)
Semantic (recursive struct check): 00:00:00.002905585 (1852.56MB)
Codegen (crystal):                 00:00:09.960081052 (2591.56MB)
Codegen (bc+obj):                  00:00:00.618658292 (2591.56MB)
Codegen (linking):                 00:00:04.913773358 (2591.56MB)

Macro runs:
 - /home/jeremy/Sites/joysticktv/lib/teeplate/src/lib/file_tree/macros/directory.cr: reused previous compilation (00:00:00.002599683)
 - /home/jeremy/Sites/joysticktv/lib/lucky/src/run_macros/generate_asset_helpers.cr: reused previous compilation (00:00:00.004127867)
 - /home/jeremy/Development/crystal/lang/src/ecr/process.cr: reused previous compilation (00:00:00.002720205)

Codegen (bc+obj):
 - all previous .o files were reused

real	0m23.521s
user	0m28.763s
sys	0m3.914s
```

So looks like it compiles with about 500mb less, and in about 1min quicker?