@HertzDevil 2022-08-19
In your tests two of the macro runs were fresh only in the first compilation.