@asterite 2022-08-20
I managed to introduce an abstraction for this concept of "zero, one or many." I had trouble doing that at first because making `push` and `concat` work would be impossible if it's a struct. So I made the choice of explicitly asking to reassign the value if it's mutated, and named the methods `with` and `without` to make that explicit.

We'll see if CI passes.

Doing that refactor actually caught a case where `@dependencies` was always created with an array of a single element, so in that case there's no allocation anymore. And that case was pretty frequent too! So that reduced the memory usage a bit more.

If all goes well, making a call args, or in general most of ASTNodes that hold an array of other nodes use this type, might give a really big memory save, and maybe also a performance boost. But I'd leave that for a later PR if we go forward with this one.