@asterite 2022-08-24
I also made another custom benchmark for this. Here's the code I used:

```crystal
require "./src/compiler/requires"
require "benchmark"

filename = ARGV[0]
code = File.read(filename)
source = Crystal::Compiler::Source.new(
  filename: filename,
  code: code,
)

Benchmark.ips do |x|
  x.report("compile") do
    compiler = Crystal::Compiler.new
    compiler.no_codegen = true
    compiler.compile(source, "dummy_output")
  end
end
```

Then you can compile that file in release mode with `master` branch or this branch and run it by passing a filename.

Some outputs:
- samples/mandelbrot2.cr
  - before: 7.32  (136.60ms) (±18.68%)  127MB/op
  - after: 8.60  (116.34ms) (±12.01%)  120MB/op
- samples/compiler/visitor_example.cr
  - before: 3.12  (320.47ms) (±22.62%)  276MB/op
  - after: 3.28  (305.28ms) (±11.20%)  252MB/op
- samples/2048.cr
  - before: 6.58  (151.94ms) (±13.18%)  134MB/op
  - after: 7.66  (130.51ms) (±15.62%)  127MB/op

So it's not a huge change, but it's definitely faster and with less memory.